requirements.txt
scikit-learn==1.8.0
mlflow==3.11.1
evidently==0.7.21
fastapi==0.136.0
uvicorn==0.44.0
pandas==2.3.3
numpy==2.4.4
matplotlib==3.10.8
seaborn==0.13.2
python-multipart==0.0.26
dvc[all]==3.67.1
docker==7.1.0
requests==2.33.1
PyYAML==6.0.3
Flask
flask-cors
types-PyYAML
joblib
ensure
tqdm
notebook
python-box